+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+## 0.0.4
+
+* Add takeScreenshotImage
+* Support scale Screenshot,
+
 ## 0.0.3
 
-* Update Dart SDK require >=3.2.0 <4.0.0
+* Fix takeScreenshot when renderObject detach
+
 ## 0.0.2
 
 * Update Dart SDK require >=3.2.0 <4.0.0